renamed to src, added directory cleanup, added docs to gitignore
