+++ hellouniverse.go
@@ -1,0 +1,21 @@
+package hellouniverse
+
+import (
+	"fmt"
+)
+
+// @@@START hellouniverse
+func HelloUniverse() {
+	fmt.Println("Hello Universe!")
+}
+// @@@END
+
+func SomeCode() {
+	fmt.Println("This is some code")
+}
+
+// @@@START hellobacon
+func HelloBacon() {
+	fmt.Println("Hello Bacon!")
+}
+// @@@END
