+++ package.json
@@ -8,7 +8,7 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "snipsync": "./snippetsync/index.js"
+    "snipsync": "./src/index.js"
   },
   "dependencies": {
     "@octokit/rest": "^18.0.3",
