+++ src/index.js
@@ -10,7 +10,6 @@ if (cfg instanceof Error) {
   logger.error("unable to read config file: " + cfg);
 }
 logger.info("config loaded:");
-logger.info(cfg);
 
 synctron = new sync.Sync(cfg, logger);
 synctron.run();
